+++ app/src/main/java/eu/kanade/domain/ui/model/AppTheme.kt
@@ -4,7 +4,8 @@ import dev.icerock.moko.resources.StringResource
 import tachiyomi.i18n.MR
 
 enum class AppTheme(val titleRes: StringResource?) {
-    DEFAULT(MR.strings.label_default),
+    DEFAULT(MR.strings.theme_sunrise),
+    CLASSIC(MR.strings.theme_classic),
     MONET(MR.strings.theme_monet),
     CATPPUCCIN(MR.strings.theme_catppuccin),
     GREEN_APPLE(MR.strings.theme_greenapple),