+++ telemetry/src/firebase/kotlin/mihon/telemetry/TelemetryConfig.kt
@@ -11,7 +11,7 @@ object TelemetryConfig {
 
     fun init(context: Context) {
         // To stop forks/test builds from polluting our data
-        if (!context.isMihonProductionApp()) return
+        if (!context.isTaisonProductionApp()) return
 
         analytics = FirebaseAnalytics.getInstance(context)
         FirebaseApp.initializeApp(context)
@@ -26,15 +26,14 @@ object TelemetryConfig {
         crashlytics?.isCrashlyticsCollectionEnabled = enabled
     }
 
-    private fun Context.isMihonProductionApp(): Boolean {
-        if (packageName !in MIHON_PACKAGES) return false
+    private fun Context.isTaisonProductionApp(): Boolean {
+        if (packageName !in TAISON_PACKAGES) return false
 
         return packageManager.getPackageInfo(packageName, SignatureFlags)
             .getCertificateFingerprints()
-            .any { it == MIHON_CERTIFICATE_FINGERPRINT }
+            .any { it == TAISON_CERTIFICATE_FINGERPRINT }
     }
 }
 
-private val MIHON_PACKAGES = hashSetOf("app.mihon", "app.mihon.debug")
-private const val MIHON_CERTIFICATE_FINGERPRINT =
-    "9A:DD:65:5A:78:E9:6C:4E:C7:A5:3E:F8:9D:CC:B5:57:CB:5D:76:74:89:FA:C5:E7:85:D6:71:A5:A7:5D:4D:A2"
+private val TAISON_PACKAGES = hashSetOf("com.gent8.taison", "com.gent8.taison.debug")
+private const val TAISON_CERTIFICATE_FINGERPRINT = ""